CSDK-1746 make participant video renderers safer
We now register an optional `on_video_frame` delegate when setting the client
delegate. Every time we register a participant video renderer we now do it with
a renderer ID and we always use the new `on_video_frame` delegate. Then, one
each participant video frame we will call the `on_video_frame` deletage with the
renderer ID.

In the case of `daily-python` we then register Python callbacks with the
renderer IDs and we set the `on_video_frame` delegate which is the same for each
participant. Finally, when a video frame arrives we look for the Python callback
using the incoming renderer ID.